chore: extract api to its own file
